renamed getPopularData to getData
